Squash housekeeping period options into a single one
With the introduction of OptimizeRepository RPC some of the
housekeeping options become redundant.

Currently, we have 3 options that are responsible for how
often a particular optimization is called:

- Incremental repack period
- Full repack period
- Git GC period

With OptimizeRepository introduced, the frequency of a particular
operation can't be controlled from Gitlab Rails anymore.
Let's choose the lowest one out of these three
(Incremental repack period) and use it as a value of how often
OptimizeRepository is called.

Changelog: changed