Implement max limit for assigned issues count
When a user has many assigned issues,
counting them can be a very slow task.

This commit implements a max limit
when counting assigned issues so that
only up to 100 assigned issues are counted
behind a new feature flag.

To support the new count query with max limit,
we are replacing an existing index for issue_assignees table.

Changelog: changed