Add the ability to get version info from derived MR info
Since we're already extracting information from the MR endpoint,
this addition also extracts the diffId and startSha if they're
available.

Many MR URLs will not have these values, but that's okay. 
Notably, the presence of the diffId is not - by itself - enough 
to indicate that the MR endpoint we're looking at is a 
"sub"-version of the whole MR. The BASE -> HEAD diff can be 
represented as a single diffId. In fact, BASE -> HEAD can also 
be represented as a combination of a diffId plus a startSha. 
So even knowing that we're for sure looking at an MR version 
isn't conceptually possible knowing only the endpoint.

Since all the endpoints in an MR lack the protocol/domain, we
use "https://gitlab.com" as a dummy root so we can successfully
construct a new URL object. Since - for this purpose - we don't
care what the root is, it's just used to get the URL constructor
to succeed. An alternative would be to use something like
`document.location.protocol` + `document.location.host`, but
this seems unnecessarily brittle given that we immediately throw
away everything except for the search parameters.