Update occurrences rather than appending new ones
The initial implementation of `sbom_occurrences` involved tracking all
the dependencies in all of the pipelines. When we enabled the feature on
production, the table grew far too quickly (almost 1 million rows per
hour). In order to allow us to ship an MVC, we're going to take a more
minimal approach. When new SBoM reports are ingested, we'll create new
records, update the existing ones, and delete the no longer present
ones. The `sbom_occurrences` table will only hold the current state of
the default branch.

Changelog: changed