Potential fix for flaky spec/features/jira_connect/branches_spec.rb
The previous test may have been racy because the test attempted to
enter `my-source-branch` in the `Search` box, but selecting the
`master` for the `Source branch` in the browser may cause the `Search`
box to be overwritten with `master`. To prevent this race, we wait for
the search request to complete before filling out another term.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/385660