Improve license text on subscription page
Prior to this change, a few texts on the subscription page could cause
confusion. An example is the "Renews: in X days" text which was
confusing to customers who disabled auto renew or have already renewed.

With this change, some texts on the subscription page as well as a text
on the admin overview page are updated to reduce customer confusion.

Changelog: other
EE: true