Restrict changing of issue type
For legacy issues it's possible to switch between issue, incident and
test_case types, but it should not be possible to switch also to other
types (which are being introduced for work items). Although we plan to
allow changing of type for work items, this is not ready yet.

Changelog: changed