Add FF around terraform state name validation
On https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105674 we made
available the use of dots in terraform state names with locked states.
The fact that we did not allow it was a bug.
But by fixing it, we have introduced a situation where other users which
used the terraform state name with a dot but with unlocked states, they
would suddenly start adding their tf state with a dot, thus losing track
of the previously existing tf state without dots. This happens because,
although dots used to work for non-locked states, they would still be
saved in the database without the dot. As everything from the dot
onwards was being stripped.

The idea of this MR is to wrap this behaviour around a FF, so that
self-managed uses don't face this problem at least until 16.0.

We'll attempt to mitigate the problem on GitLab.com by communicating to
our users that they should migrate their TF_STATE_NAMES

This also adds a migration to enable the FF immedieately on Gitlab.com.