Fix time-dependent test in spec/controllers/projects_controller_spec.rb
This test may have been failing intermittently when the time
approaches a minute interval, which causes the rate limit to expire,
and allows the project export to proceed. Fix this by freezing the
time so that this test is agnostic to the current time.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/382028