Increase max limit on url_variables token length
Drone CI token URI variable is an Oauth2 token.

There is no set maximum length for what an OAuth token could be.

Google API seems to have the longest I've found, saying it can be up to
2048 bytes.

We can drop the maxLength entirely from the JSON schema to validate, but
it's probably sensible to come up with a max limit unless we're happy to
save any length of value (which we probably aren't).

This change sets the new limit to 2048 characters, based on the high
bytesize given of an Google API token. Tokens are likely to be ASCII.

https://gitlab.com/gitlab-org/gitlab/-/issues/385919#note_1211528329

Changelog: fixed