Enable pages namespace domains in dev mode
Gitlab Pages supports namespace domains, like `user.gitlab.io` or
`group.gitlab.io`, documented in
https://docs.gitlab.com/ee/user/project/pages/getting_started_part_one.html

To do that, the pages project must match the domain completly. For
example, to have `group.gitlab.io` one needs to great a project
`group/group.gitlab.io`.

But, this wasn't supported when running Gitlab on a non 80 port, like in
GDK where we run Gitlab on the port 3000.

This commit enable developers, in dev mode, to use this feature when the
port is not 80.