Fix update templates Rake task for development environment
Now that Gitaly Cluster manages repository directories, the update
templates Rake task failed because it attempted to modify a Git
repository directly on disk, which did not exist in the `@hashed`
directory path.

To fix this, we now push directly to the SSH remote. This requires the
admin user to have an SSH key and the gitlab-sshd running in the GDK.

Changelog: fixed