Fix spec that was dependent on GitLab.com state
The submodule test in `spec/features/projects_spec.rb` clicked on the
link to the `gitlab-grack` submodule, which directed the browser to a
repository on GitLab.com. The test then proceeded to check the
reference selector of GitLab.com, but this can change depending on the
version deployed.

To avoid testing against GitLab.com, adjust the test to ensure that a
link is available pointing to the right location.

Relates to
https://gitlab.com/gitlab-org/quality/engineering-productivity/master-broken-incidents/-/issues/412