Switch to use a button click to start a new thread
The UX here is essentially identical to the previous version:
Both use a click to start leaving a comment. The previous version
used a click inside an input, and this one uses a click on a button.

The - important - difference is that the "new thread" interaction is
more visually distinct from the "reply" input on existing threads
above in the discussion tree.

Changelog: changed