ci: Build Workhorse with FIPS Go compiler
To test Workhorse against the FIPS Go complier, we need to use a
Debian-based image since we need to install exiftool. We also need to
define the `FIPS_MODE` environment variable so that the `fips` build
tag is added. Omnibus GitLab already ships an image that we can use
to verify Workhorse runs correctly in this environment.

Changelog: changed