Allow null `only`/`except` keywords to be combined with `rules`
Allow the keywords `only`, `except`, and `rules` to be combined
provided either `only`/`except` is (or are both) null or `rules`
is null. This enables users to extend an existing job that
uses `only`/`except` and override it by setting it to null.
Similarly, `rules` can be set to null. Note that the yaml value
for null is either: null, ~, or <blank>.

Changelog: changed