Allow member creator service to handle existing access requests
If a user requests access to a group prior to SAML Group Sync
subsequently attempting to add the user as a member, the
process would fail with access denied. This change skips
authorization on acceptance of access requests when the action
occurs without an acting user. This is consistent with the approach
to adding members in general.

Changelog: fixed