Pass --strip-component when running tar -zxf
The wrapping directory inside the archive is usually called
`${ref}-${sha}-${path}`, which the previous `mv` does work
because `ref` will match. However when `ref` contains `/`,
it'll be translated to `-` which does not match.

Instead of trying to match it, use `--strip-component` can avoid
this and we don't need to use `mv`.