Relax FIPS constraints on PyPi packages
Previously on a FIPS system if a PyPi package were uploaded that
included `md5_digest`, the upload would fail with a 422 Unprocessible
Entity error due to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87180.  This
commit relaxes the constraint and only fails if `sha256_digest` is not
present.

Even on a FIPS system, running a Docker image such as `python:latest`
doesn't ship with a FIPS-enabled OpenSSL. The change in
https://github.com/pypa/twine/issues/776 doesn't omit `md5_digest` as
a result.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/385477

Changelog: fixed