Refactor Gitaly client stub to reduce the number of connections
Recently, GitLab Rails cache the client stubs by (storage, service).
When creating a new Client object in Ruby, it dials and maintains a
connection to the target host. It means a GitLab Rails process may open
too many connections to Gitaly cluster. The client opens up to hundreds
of connections per host. It's a waste of resources because all clients
in the same host can share the same connection. It's the purpose of
HTTP2 transportation from the first place. Each connection may only be
used a couple of times in a life cycle of the Rails process. In
addition, establishing a new connection has overhead. It may add
hundreds of latency before any call.

This commit creates and caches gRPC channels by address and credentials.

Changelog: other
Signed-off-by: Quang-Minh Nguyen <qmnguyen@gitlab.com>