Add release manager link on security template
Given the broad responsibility of release managers across the different
release types, they should not review security merge requests. This is
part of an effort to limit the responsibilities of a release managers,
and also have a diverse set of eyes on these merge requests.

This commit updates the security merge request template to link to the
release managers' schedule.