Drop injects' default values
This removes default values from injected properties within the
`NotificationEmailListboxInput` component. Those were converted from
props to injects late in the initial implementation and the default
values had been preserved. However, it seems like an injected prop only
falls back to its default value if the property isn't provided at all.
If the property is provided but is `undefined`, that's used as the
prop's value, which caused an issue with the input's default value which
needs to be an empty string and not `undefined`.

Changelog: fixed