Drop justify-content-center class in favor of align prop
With some recent markup changes in `GlPagination`, the
`table-pagination` component's alignment doesn't seem to apply properly
anymore. The reason likely is that alignment classes need to be set on
the `<ul>` tag, which is now wrapped within a `<nav>` tag.
Instead of manually adding the `justify-content-center` class, we can
leverage the `align` prop which will take care of adding the class to
the correct element.

Changelog: fixed