Trigger merge status subscription when a new discussion is created
When a new unresolved discussion is created (e.g. via a review or
starting a new thread), it affects the mergeability of the merge
request.

Since we need to update the merge widget in real-time, we need to
let the subscribers know about it. To do this, we trigger the
subscription whenever a new unresolved discussion is created.

Changelog: added