Enforce memory-watchdog by default
We have fully rolled this out so we don't need
the outer FF anymore, plus, we should enforce
it by default since it replaces PWK for Omnibus
users, which was also enabled by default.

This fixes the problem that no memory killer was
running anymore after the swap from PWK to the
watchdog in 15.6.

Changelog: fixed