Restructure subheading to lead with error msg
The information in the previous commit was good - it includes the
actual error messages for searching purposes! - but it needed
blending with the rest of the subheading. It feels appropriate to
move the error messages to the top of the subheading, so users know
immediately whether or not this section applies to them. I'm hesitant
to edit the subheading, because deep links. I think it can stay
what it is for now.