Fix bug in backfilling Container Repository
The `RepositoryRegistrySyncWorker` uses
`container_repositories_max_capacity` because it
is intended to schedule container repository sync
jobs for backfilling, retrying failures, and to
cover any rare lost events.

However it is not handling container repositories,
and is only handling Git repositories. This makes
sense long-term. But we need to force it to handle
container repositories as well for now until we
have https://gitlab.com/gitlab-org/gitlab/-/issues/372444#note_1087132645c.

No changelog needed since this fix is occurring
before the bug was released.