Fix suggested reviewers job not run on branch MR
This fixes a regression issue where suggested reviewers job not running
on merge request created from a branch. The `modified_paths` check for
`FetchSuggestedReviewersWorker` depends on `merge_request_diff` being
persisted as part of `NewMergeRequestWorker`. However, in a recent
refactor, `NewMergeRequestWorker` is scheduled to run after the check is
invoked and hence causes this bug.