Show new "Your work" sidebar on Snippets pages
This adds a new menu item "Snippets to the new "Your work" sidebar and
shows the sidebar on all snippets pages, like new, edit, show, and the
dashboard itself.

The sidebar is not shown while exploring or showing public snippets of
other users, as this belongs into the "Explore" section which will get
its own sidebar.