Fix manual variable inputs width
When inputting manual variables value, ther ewidth of the input
value field would change to accomodate the new clear button
being added to the DOM. To prevent this issue, we add an invisible
button that has no click events which is always in the layout.

We are also adding tests relating to this change.

Changelog: fixed