Fix 404 when clicking 'Cancel' from fork project page
Currently clicking 'Cancel' from the 'Fork Project' page causes a 404.

This is because the href on the 'Cancel' button back to the project
page is considered a relative link because it does not begin with '/'
and so resolves to something like
/namespace1/project1/-/forks/namespace1/project1 instead of
/namespace1/project1.

This change resolves this by providing `cancelPath` to the cancel button
component using the Rails path helper for the project instead of
`Project#full_path`.  The Rails path helper prefixes the path with '/',
`Project#full_path` doesn't.

Ref https://gitlab.com/gitlab-org/gitlab/-/issues/378502

Changelog: fixed