Shifts branch rules info location, tidies up
At this time, I don't think it's a good idea to make a separate page
for this information. Potentially, eventually it will be the right
decision, but right now it is not.

This commit takes the information that Nataliia and Joe contributed,
and moves it to an existing page. Deleting the page they contributed to
removes their names from the commit history (boooo!) so I'm crediting
them as co-authors here. The bulk of the information came from Nataliia,
with Joe contributing wording changes. I layered my edits on top.

Teamwork FTW.

Co-authored-by: Nataliia Radina <nradina@gitlab.com>
Co-authored-by: Joe Woodward <jwoodward@gitlab.com>