Downplays feature flag, plays up warning
Don't treat this flag like an ordinary feature flag, because it is
not. Make the instructions smaller. Move them to the bottom of the
page in the troubleshooting section. Make the warning as dire as
possible, clarify that you shouldn't do this unless you know what
you're doing, and try to downplay its existence as much as possible.