Make dropdown menu item with nav toggle clickable
Now a click on the dropdown menu item, outside of the toggle itself,
will also switch the toggle and reload the page with the other nav.

Also, clicks on that menu item (or the toggle) don't close the
dropdown menu anymore. This allows the toggle animation to play before
the page gets reloaded.

Keyboard navigation ignores the menu item but focuses only the toggle.