Add zoekt services to run in CI
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105049 we added
a new Zoekt integration for code search. This MR introduced a few RSpec
tests with the `:zoekt` tag that we needed to skip in CI because they
depend on having running zoekt services. We have since created a docker
image in
https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/624
for running the necessary dependencies in CI as `services`. This MR
adds this image as a service and removes the rule for ignoring the
`:zoekt` tag in CI.

We also needed to update Webmock stubbing to allow requests to these
services.