Fix pages internal cache
A gitlab pages project might have multiple domains, like:
- domain.com
- domain.co.uk

While enabling the `cache_pages_domain_api` Feature flag, customers with
gitlab pages projects that had multiple custom domains complained that
their sites stop working.

This was happening because while caching the `/api/internal/pages`
payload, we were using the `:project_id` in the cache key, which was
caching the first request for the first domain in the payload. For
example, if the first request was for `domain.co.uk`, even when
requesting `domain.com` we'd return the `domain.co.uk` data, which
includes the https certicates.

To fix that, instead of the `project_id`, the cache for custom domains
are using the `domain_id`. This each domain will have it's own cache.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/385865
Changelog: fixed