Update flaky failure template to account for stable branches
Delivery is working to extend the GitLab maintenance policy to support
bug fixes to the two previous monthly releases in addition to the
current stable release (details on
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/828). Part of this
work will open up the stable branches to allow developers to merge bug
fixes directly into the stable branch.

Stable branches represent the source of GitLab releases, and because of
this it is vital to guarantee their readiness and availability.

This commit expands the flaky-failure template to encourage the fix to
be cherry-picked into the stable branches, which is already part of the
broken-master resolution.