Trigger merge status subscription when blocking MR gets added/removed
When a blocking merge request is added/removed to a merge request,
it affects the mergeability of the merge request.

Since we need to update the merge widget in real-time, we need to
let the subscribers know about it. To do this, we trigger the
subscription whenever that happens.

Changelog: added
EE: true