Create indexing-restrictions component
This creates the `indexing-restrictions` component which will replace
the current `select2`-based selector in the admin ElasticSearch indexing
restrictions settings.

This new component leverages `GlCollapsibleListbox` to fetch and present
available index entities. Selected entities are displayed in a
scrollable list below the selector and can be removed from there. A
hidden input is kept in sync with the selection.

For the sake of keeping the changes small, this only introduces the
components and their tests. We'll actually use them in a follow-up.