Make new-issue-dropdown more reusable
This moves the `new-issue-dropdown` component to the `vue_shared/`
directory and changes it slightly to make it more reusable. Most
noteably, the Apollo query is made configurable so that we can retrieve
projects from outside of a given group.
For now, these changes aren't being leveraged, but they will be in
follow-ups.