Expand chain-of-custody report from merge commits to all commits
[Issue 267601][0] requests to expand the chain of custody report from
Merge Commits to All Commits.

The meaning of "all" is a little bit vague, so I will expand on the
implemntation below.

First thing to mention is that this change is feature flagged with the
`all_commits_compliance_report` flag. This will allow us to iterate on
the implementation with input and feedback from production use.

Background
==============
The compliance report prior to this change is a CSV report that
contains info from all Merge Requests under a given group, in the past
month.

The report is generated in a worker, and emailed to the requesting
user. There is a 15mb max limit on the CSV due to the fact it is being
emailed.

Implementation
======================
Since there is a 15mb limit, it doesn't really make sense to
grab *all* commits, especially for a group with a lot of active
projects, as they will ultimately be truncated.

The current strategy I took is:

1. Get all projects under a group
2. Iterate over all projects, grabbing all repository commits
  2a. I am only grabbing from a 1-month window
  2b. I am only grabbing `COMMITS_PER_PROJECT` max commits
3. combine all of those commits into a collection
4. feed them into the CSV builder, letting it truncate at 15mb
   wherever that falls at

The existing "merge commit filter" feature
=============================================
There is a feature to generate this report for a specific MR sha. That
feature is still "MR Only", but will also be expanded to all commits
once the new version of the report has baked in production a bit

Getting a merge request from a gitly commit
================================================
This feature is supposed to be "expanding" the existing report, not
replacing it. As such, it seems like we should keep the existing
columns and polpulate them whenever a `gitly` commit can be mapped to
an MR

The only way I really found to do that is using the existing MR
finder while iterating iver every commit. This seems "not very
efficent" to me, but I couldn't find another way to map from gitly
commits to MRs

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/267601

Supporting changes
--------------------

CommitCollection
==================

In order to be able to pass a [`CommitCollection`][1] to the
[`CsvBuilder`][2] library, there needs to be a `find_each` method
defined. Since repository commits are not actually active record
object, here we just define a `find_each` method that aliases the
`CommitCollection#each` method

[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/37f1faa857b2dd7492f981f25609fd1167ce3070/app/models/commit_collection.rb
[2]:https://gitlab.com/gitlab-org/gitlab/-/blob/37f1faa857b2dd7492f981f25609fd1167ce3070/lib/csv_builder.rb

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/102364
EE: true