Fix flaky merge request widget spec
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/108123 added
a merge request widget that shows "Security scans have run" even
when there are no actual security scans present.

This new widget is behind a feature flag, so let's just disable this
flag in the test until the feature is fully flushed out. This is
already done in an EE spec.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/386691