Stop using hideFlash
For `vue_merge_request_widget` related files, we replace `hideFlash` by
keeping track of created alerts and use their `dismiss()` methods.

For the rest, we replaced `hideFlash` by clicking on the close button in
the alerts (JavaScript) or flash (Rails) elements.

On a follow-up we will remove the `hideFlash` code altogether. We can't
do it now, because `addDismissFlashClickListener` depends on it.