Trigger merge status subscription when blocking MR gets merged
When a blocking merge request is merged, it affects the mergeability
of the merge request/s it was blocking.

Since we need to update the merge widget in real-time, we need to
let the subscribers know about it. To do this, we trigger the
subscription whenever that happens.

Changelog: added
EE: true