Rename type_id argument to prevent confusion
The 'type_id' argument is renamed to `type_iid` in this commit.

Despite the name, the argument stores iids of objects and has led
to incorrectly specified tests that end up passing ids of objects
to QuickActions::TargetService.

Many examples in 'autocomplete_sources_controller_spec.rb'
passed the ids of issues to the project autocomplete controller
which, in turn, passed along the ids to "QuickActions::TargetService".
The tests have still passed because the service builds temporary
objects using Builders. This commit also removes the `target_id` params
from the tests where they weren't needed in the first place.