Propagate instrumentation params for etag-hits
This includes the expected information in the ActiveSupport
notification that populates the payload for logging by
lograge. Specifically, adds the information we pushed into the context
for etag-caching to the logging message.

We also add the urgency information to the env so it can be picked up
by RackRequestsMiddleware to emit the correct urgency labelling.

Changelog: fixed