Cleanup project_select_button conditions
This removes the code associated to the `project_select_button` local
assign in a few views. In the event `project_select_button` evaluated to
`true`, we would render a project selector in a couple empty states.
However, `project_select_button` was never actually `true`, so there
is no reason to keep those conditions in the code.