Upgrade sprockets gem to 4.1.1
Changelog: changed

Introduce manifest.js file for handling precompiled assets

This manifest contains page bundles and non-standards assets
that needs to be precompiled. Before we were adding them to
application.rb but with sprockets upgrade we have to use manifest.js.
For EE and JH there is no corresponding manifest.js file
as this is not supported by sprockets.

Instead for EE and JH we use manifest-like files ee.js and jh.js
located under the "ee/app/assets/config/" directory.
For JH the directory is "jh/app/assets/config/".