Re-schedule epic cache update
We should re-try cache update for all epics for which we couldn't obtain
an exclusive lease. The reason is that if an epic/issue is moved in the
same epic  hierarchy then cache update can be scheduled multiple times
and we should make sure that "roll-up" scheduled job is executed.