Decouple UI concerns from group-select component
This refactors the `group-select` component by extracting its user
inteface concerns into a new data-agnostic component (`entity-select`).
This will let us create more flavors of selectors without having to
reinvent the wheel as we can let the base component do the heavy-lifting
for us, and focus on the fetching and error handling in the
entity-specific components.