Replace .latest scope with .merge
The Security::Finding.latest scope includes a `join` call on
`security_scans`.  Using this in conjunction with `.with_scan` - as is
done in this method - causes the `security_scans` table to be joined
twice.

We can use `.merge` here and we do not need the extra `.join` call as
`security_scans` is already included in the query from `.with_scan`.

Fixes 384102