Set Cache-Control max-age when ck param is present
To utilize client side caching, we are adding support for `ck`
param (short for cache key). When that param is present, the
`Cache-Control` header will be set with a `max-age` of `86400`
which is equivalent to a day. This will tell the browser to cache
the request for a day.

There's no effect right now as ~frontend will need to pass the
`ck` param for this to work.

This is behind the `diffs_batch_cache_with_max_age` feature flag.