Refactor invalid scope
Rails scopes will return an ActiveRecord::Relation if nil is returned.
Using find_by in a scope can make it behave like Klass#all if nil is
returned. e.g.

  class LfsObject
    scope :valid, -> { find_by(id: first.id) }
    scope :invalid, -> { find_by(id: nil) }
  end

  > LfsObject.valid == LfsObject.first
  => true
  > LfsObject.invalid == LfsObject.all
  => true

It's unlikely the scope in question caused any issues. It is only
referenced in one location and we call #id on it which would result in a
NoMethodError