Fix `[nil]` arrays
A `[nil]` array can come from the `FindingsFinder` for remediations.
This causes an error state with GraphQL when it converts it to `nil` (no
array). The `PipelineVulnerabilitiesFinder` also can produce `[nil]`
array in the right circumstances, so this also protects against that if
we add fields later.

Changelog: fixed
EE: true