Switch to upstream omniauth_openid_connect gem
We forked the original gem a while ago into to fix a number of issues
and add a number of features.

Since then we've upstreamed all the changes into the
omniauth_openid_connect repository. In addition, the upstream project
has added PKCE support and has other contributors.

This commit locks the openid_connect gem to v1.3.0 since upgrading
past that version pulls in an updated net-smtp, which cannot be used
with Ruby 2.7 due to https://bugs.ruby-lang.org/issues/17761.  See
https://docs.gitlab.com/ee/development/emails.html#rationale for more
details.

Relates to
https://gitlab.com/gitlab-org/ruby/gems/gitlab-omniauth-openid-connect/-/issues/5

Changelog: changed