Bump hard stop version, auto-detect min schema
We discovered that there are variables used by schema check before doing
a migration which must be manually updated every time a migration squash
happens.

This change automates detecting the minimum schema version acceptable to
perform a migration to bring the database up to date. The minimum GitLab
version is still hardcoded because this value should only be adjusted
when an intentional hard stop is introduced.