Consolidate discussion GID for user comments
When a user discussion contains a single note the discussion GID is
different from when a user discussion contains more than one note.

It is different because in the first place discussion is instanced
with IndividualNoteDiscussion whereas in the second case it is instanced
with Discussion.

This fixes that so that user discussion always returns the same
discussion GID regardles of the number of notes in the discussion.