Assign ArkoseLabs risk band to users that signed up with OAuth
When a new user is created using an external OAuth provider (e.g. sign
in/up with Google), we redirect them to a page where they need to solve
the ArkoseLabs challenge.

Solving the challenge will assign an ArkoseLabs risk band to the user
which is used to determine what verification methods (credit card /
phone number / email) will be required for them as part of the identity
verification process.

Changelog: added
EE: true