Calculate and use sha for CI config variables cache
Today, we are receiving "ref" from client and calculate config variables
according to it. We also use this "ref" to decide the cache key.
However, "ref" is not a good candidate for a cache key because the name
is static but that branch can be changed.

In this commit, we are calculating and using "sha" of the "ref" for
the cache key.

Here, we are also renaming the "sha" parameter to "ref" because we are
actually receiving "ref" from client.

Changelog: fixed