Fix incorrect child computation
This change fixes a UI bug which occurred when a widget renders with an
empty wrapper component. The parent component would check if there is any
child component, and showed the border of the report section accordingly
which resulted in a glitch that showed a thin border for empty wrappers.
Instead of checking for a tag, checking if the child has any text
content fixes this bug. If there is no text, let's not show the component.

Changed: fixed