Show assigned and reviewer MRs as flyout menu items
Before, the "Merge Requests" link in the new "Your work" sidebar showed
only the count of assigned merge requests.
Now it shows the sum of assigned and reviewer MRs and shows both
seperate as flyout menu items.

For this, the menu_item.rb had to be extended to also be able to show a
count.