Render Jira app views based on key storage enabled
If the jira_connect.enable_public_keys_storage config is enabled
the GitLab instance can serve as a proxy for GitLab for Jira app
installation and should render the GitLab.com view.

Usually this is only the case for GitLab.com. This fixes a bug
where self-managed users see an GitLab.com option when manually
installing the GitLab for Jira app.

Changelog: fixed