Moves info from API page to new changelog page
Cuts a large section of information out of doc/api/repositories.md
and moves it to a new page, doc/user/project/changelog.md. The info
will need a LOT of revision to get it ready to go in user
documentation, but it was never gonna get attention buried in an
API doc like it was.