Implement batch request for CI project includes
As we did this for local includes, now we are implementing the same
approach for project includes.

Our main goal is to minimize the number of call to Gitaly. In this
commit, we are collecting the Gitaly calls via BatchLoader and making
a single Repository#blobs_at call for all project files.

This change is behind FF ci_batch_request_for_local_and_project_includes