Sync scan result policies when project assigned to a group
This commit fixes a bug when a project is created/transferred
to a group, the scan result policies are not synced. This fix
will call the worker to sync the scan result policies when a
project is created inside a group. If a project is transferred
from one group to another, existing security policy project
will be unassigned.

EE: true
Changelog: fixed