Second round of structural revisions
Bring this page closer to GitLab tone and style. Revise subheading
nesting. Rewrite a LOT of text for brevity, tone, and style. Move
code examples in ordered lists (some of them, anyway) into inline
code to make the flow of the instructions easier to read.

Also, rename from "force-push" to "force push" because it looks
closer to what the Git documentation uses.