Add zoekt code search integration
This [MR](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105049)
introduces a new backend [Zoekt](https://github.com/sourcegraph/zoekt)
to power code searches in GitLab. Previously we supported Elasticsearch
and Gitaly backed searches. We found Gitaly didn't scale for searching
across large groups and Elasticsearch has limitations in how indexing
and searching works that makes it a poor choice for code. Zoekt provides
a more "grep-like" user experience and solves a lot of the problems we
had with Elasticsearch. You can read more about the motivation at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/107891 . Since the
rollout of this will be gradual and it is somewhat experimental it will
not be the default and we will not be removing support for Elasticsearch
any time soon.

Changelog: added