Add Geo::WikiRepositoryState model
This changes adds a new model to fix the wrong primary
key defined in the Geo::ProjectWikiRepositoryState after
the introduction of the Projects::WikiRepository model.

We are adding a completely new model and table. This way
we can drop the table and remove the old model at a later
step without affecting the release date. We can also
rename the model and table later without downtime.

Changelog: added